refactor(driver): move sort prune to qlang
