+++ jina/drivers/querylang/sort.py
@@ -1,7 +1,7 @@
 __copyright__ = "Copyright (c) 2020 Jina AI Limited. All rights reserved."
 __license__ = "Apache-2.0"
 
-from . import BaseDriver
+from .. import BaseDriver
 
 
 class TopKFilterDriver(BaseDriver):
